Server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
pytest
pytest-asyncio
httpx
aiosqlite
